Adding an example application for Mac which uses the model layer defined in the Core framework.
This application, which I wrote in about 15 minutes includes the BWToolkit, but would advise using a git submodule for the framework source. Note that the only class in this project is the App Delegate, which only includes a property for a managed object context, which we just get from the MOCManager. All the other functionality (which isn't much) is done in IB using NSArrayControllers and Cocoa bindings.

Signed-off-by: Daniel Thorpe <dan@blindingskies.com>